now it can show what is being played, and presents a terminal as output.

+++ lastfm-scrobler.py
@@ -5,10 +5,12 @@ import pylast
 import os
 import urllib2
 import re
-import threading
 import ConfigParser
+import time
 
 network = None
+now_playing = None
+now_playing_countdown = 20
 
 def get_config(config_file):
 	config = ConfigParser.ConfigParser()
@@ -26,14 +28,10 @@ def getLastFMScrobbleObject():
 	network = pylast.LastFMNetwork(api_key = API_KEY, api_secret = 
 	        API_SECRET, username = username, password_hash = password_hash)
 
-	#track = network.get_track("石进", "初恋的美好")
-	#print track.get_duration()
-	#print track.get_playcount()
-
 	return network
 	
 def parse_song_info(song_id):
-	#print song_id
+
 	xml_loc = 'http://www.xiami.com/song/playlist/id/' + song_id + '/object_name/default/object_id/0'
 	#print xml_loc
 
@@ -52,7 +50,7 @@ def parse_song_info(song_id):
 		#print song_title
 	else:
 		print 'song_title not match'
-		song_title = "未知"
+		song_title = "Unknown Song"
 
 	artist_reg=r"<artist><\!\[CDATA\[(.*)\]\]></artist>"
 	match = re.search(artist_reg, text)
@@ -61,37 +59,72 @@ def parse_song_info(song_id):
 		#print artist
 	else:
 		print 'artist not match'
-		artist = "未知"
+		artist = "Unkown Artist"
 
-	return (song_title, artist)
+	return (song_title, artist.split(";")[0])
 
 
-def handle_cached_dir(dir):
-	print "start handle cached history"
-	global network 
+def handle_cached_dir():
+        dir = "data"
+	global network, now_playing, now_playing_countdown
+
+        now_playing_countdown = now_playing_countdown - 1
+
 	filelist = os.listdir(dir)
+        count = len(filelist)
 	for onefile in filelist:
 		filepath = os.path.join(dir,onefile)
+
+                count = count -1
 		if os.path.isfile(filepath): 
 			#filename: timestamp-songid
 			info = onefile.split('-')
-			print info[0], info[1]
-			(song_title, artist) = parse_song_info(info[1])
-			print song_title, artist
-			#network = getLastFMScrobbleObject()
-			network.scrobble(artist, song_title, info[0])
-			os.remove(filepath)
-	timer_start()
+			
+                        if count == 0:
+                                if now_playing != onefile or now_playing_countdown == 0:
 
-def timer_start(): 
-	t = threading.Timer(60, handle_cached_dir, {"data"}) 
-	t.start() 
+                                        (song_title, artist) = parse_song_info(info[1])
+                                        network.update_now_playing(artist, song_title, info[0])
+                                        if now_playing_countdown > 0:
+                                                print "Now playing:", artist, "-", song_title
+                                        now_playing = onefile
+                                        now_playing_countdown = 20
+
+                        else:
+                                (song_title, artist) = parse_song_info(info[1])
+                                if int(time.time())-int(info[0]) > 30:
+
+                                        network.scrobble(artist, song_title, info[0])
+                                        print "Scrobbled: ", artist, "-", song_title
+                                else:
+                                        print "Skipped: ", artist, "-", song_title
+
+                                os.remove(filepath)
+
+def handle_fav_dir():
+	global network
+        dir = "fav"
+        filelist = os.listdir(dir)
+
+        for onefile in filelist:
+                filepath = os.path.join(dir,onefile)
+
+                if os.path.isfile(filepath): 
+                        (song_title, artist) = parse_song_info(onefile)
+                        print "Added to favorites:", (song_title, artist)
+                        track = network.get_track(artist, song_title)
+                        track.love()
+
+                os.remove(filepath)
 
 def main():
 	global network
 	network = getLastFMScrobbleObject()
-	print "start works!"
-	timer_start()
+	print "Scrobbler initiated."
+        while (True):
+                handle_cached_dir()
+                handle_fav_dir()
+                time.sleep(1)
 
 if __name__ == '__main__' :
-	main()
+        main()
